Created code to compare fmw between source and destination servers

+++ FMEServerJob.py
@@ -7,10 +7,10 @@ from ApiException import APIException
 
 class FMEServerJob(object):
 
-    def __init__(self, app_config, key):
+    def __init__(self, app_config, secrect_config, key):
         self.key = key
         self.app_config = app_config
-        self.job = FMEAPI.FmeApis(app_config, key + "_fme_server", app_config[key + "_token"])
+        self.job = FMEAPI.FmeApis(app_config, key + "_fme_server", secrect_config[key + "_token"])
         ok = self.job.check_health()
         if not ok:
             raise Exception("FME Server is not healthy.")
@@ -36,21 +36,40 @@ class FMEServerJob(object):
         return fmw_list
 
     def create_repo(self, repo):
-        try:
-            self.job.get_repo_info(repo)
+        if self.repo_exists(repo["name"]):
             return
-        except Exception:
-            pass
         self.job.create_repo(repo)
+
+    def repo_exists(self, repo_name):
+        try:
+            self.job.get_repo_info(repo_name)
+            return True
+        except APIException:
+            return False
 
     def delete_repo(self, repo_name):
         self.job.delete_repo(repo_name)
+
+    def delete_fmw(self, repo_name, fmw_name):
+        self.job.delete_fmw(repo_name, fmw_name)
+
+    def fmw_exists(self, repo_name, fmw_name):
+        try:
+            self.job.get_repo_fmw(repo_name, fmw_name)
+            return True
+        except APIException:
+            return False
+
+    def get_repo_fmw(self, repo_name, fmw_name):
+        fmw = self.job.get_repo_fmw(repo_name, fmw_name)
+        return fmw
 
     def download_fmw(self, repo_name, fmw_name, file_path, overwrite=False):
         self.job.download_fmw(repo_name, fmw_name, file_path, overwrite)
 
     def delete_fmw(self, repo_name, fmw_name):
-        self.job.delete_fmw(repo_name, fmw_name)
+        if self.fmw_exists(repo_name, fmw_name):
+            self.job.delete_fmw(repo_name, fmw_name)
 
     def download_repo(self, repo_name, dir):
         fmw_dir = os.path.join(dir, repo_name)
@@ -64,10 +83,16 @@ class FMEServerJob(object):
         fmw_dir = os.path.join(dir, repo_name)
         if not os.path.exists(fmw_dir):
             raise Exception("File not found: %s" % fmw_name)
-        if not overwrite:
+        if overwrite:
             try:
-                self.job.get_repo_fmw(repo_name, fmw_name)
-                return
-            except APIException:
+                self.job.delete_fmw(repo_name, fmw_name)
+            except:
                 pass
-        self.job.upload_fmw(repo_name, fmw_name, fmw_dir)
+            if self.fmw_exists(repo_name, fmw_name):
+                raise APIException("Failed to transfer %s\%s" % (repo_name, fmw_name))
+        else:
+            if self.fmw_exists(repo_name, fmw_name):
+                return
+        self.job.upload_fmw(repo_name, fmw_name, dir)
+        if not self.fmw_exists(repo_name, fmw_name):
+            raise APIException("Failed to transfer %s\%s" % (repo_name, fmw_name))
